+++ CHANGES.rst
@@ -7,6 +7,8 @@ CHANGES
 - Pass algorithm explicit to ``jwt.decode()`` to avoid some vulnerabilities.
   For details see the blog post by Tim McLean about some
   "`Critical vulnerabilities in JSON Web Token libraries`_".
+- Allow expiration_delta and leeway as number of seconds in addition to 
+  datetime.timedelta.
 
 .. _Critical vulnerabilities in JSON Web Token libraries:
   https://www.chosenplaintext.ca/2015/03/31/jwt-algorithm-confusion.html